fix: fix search function bug
Before fix, search function would call buildVenues function to update the venueIndex
for the updated filtered list. However, venueIndex would not be reset when the search
bar was cleared, resulting in some events having the wrong venueIndex. Now, the buildVenues
function is called once the search bar has been cleared in order to reset the venueIndex.