feat: optimize infoWindow content
Previously, content for the info window was recreated each time a map marker
was activated - the content string had to be parsed into HTML, and then the
Knockout data-bindings had to be applied. This created a performance bottleneck.

Now, the template content is applied once upon the app's launch. Note: when the
info window is closed, the content must be temporarily stored in the DOM and
re-applied to the info window when it is opened again.